fix: update all test imports after private module refactoring
  This commit completes the major library hygiene refactoring by fixing
  all broken import paths in the test suite after renaming modules to
  use leading underscore convention.

  Changes:
  - Fixed 32+ test files with broken imports after module renaming
  - Updated imports from sql_testing_library.core → sql_testing_library._core
  - Updated imports from sql_testing_library.adapters → sql_testing_library._adapters
  - Updated imports from sql_testing_library.pytest_plugin → sql_testing_library._pytest_plugin
  - Fixed original_decorator_instance scope issues in 8 database test files
  - Updated patch decorator paths in test_core_extra.py
