+++ pyproject.toml
@@ -59,7 +59,7 @@ pre-commit = ">=3.5.0"
 commitizen = ">=3.0.0"
 
 [tool.poetry.plugins.pytest11]
-sql_testing = "sql_testing_library.pytest_plugin"
+sql_testing = "sql_testing_library._pytest_plugin"
 
 [tool.mypy]
 python_version = "3.9"
