+++ src/sql_testing_library/_adapters/athena.py
@@ -9,8 +9,8 @@ from typing import Any, List, Optional, Type, Union, get_args
 import boto3
 import pandas as pd
 
-from ..mock_table import BaseMockTable
-from ..types import BaseTypeConverter
+from .._mock_table import BaseMockTable
+from .._types import BaseTypeConverter
 from .base import DatabaseAdapter
 
 
@@ -145,7 +145,7 @@ class AthenaAdapter(DatabaseAdapter):
 
     def format_value_for_cte(self, value: Any, column_type: type) -> str:
         """Format value for Athena CTE VALUES clause."""
-        from ..sql_utils import format_sql_value
+        from .._sql_utils import format_sql_value
 
         return format_sql_value(value, column_type, dialect="athena")
