+++ src/sql_testing_library/_adapters/bigquery.py
@@ -8,8 +8,8 @@ from typing import Any, List, Optional, Type, Union, get_args
 import pandas as pd
 from google.cloud import bigquery
 
-from ..mock_table import BaseMockTable
-from ..types import BaseTypeConverter
+from .._mock_table import BaseMockTable
+from .._types import BaseTypeConverter
 from .base import DatabaseAdapter
 
 
@@ -97,7 +97,7 @@ class BigQueryAdapter(DatabaseAdapter):
 
     def format_value_for_cte(self, value: Any, column_type: type) -> str:
         """Format value for BigQuery CTE VALUES clause."""
-        from ..sql_utils import format_sql_value
+        from .._sql_utils import format_sql_value
 
         return format_sql_value(value, column_type, dialect="bigquery")
