+++ src/sql_testing_library/_exceptions.py
