+++ tests/test_bigquery_adapter.py
@@ -8,8 +8,8 @@ from unittest import mock
 
 import pandas as pd
 
-from sql_testing_library.adapters.bigquery import BigQueryAdapter, BigQueryTypeConverter
-from sql_testing_library.mock_table import BaseMockTable
+from sql_testing_library._adapters.bigquery import BigQueryAdapter, BigQueryTypeConverter
+from sql_testing_library._mock_table import BaseMockTable
 
 
 # Mock google.cloud.bigquery for testing
