+++ tests/test_pydantic_models_integration.py
@@ -10,7 +10,7 @@ import pytest
 from pydantic import BaseModel, Field
 
 from sql_testing_library import TestCase, sql_test
-from sql_testing_library.mock_table import BaseMockTable
+from sql_testing_library._mock_table import BaseMockTable
 
 
 class UserInput(BaseModel):
