+++ tests/test_redshift_adapter.py
@@ -8,8 +8,8 @@ from unittest import mock
 
 import pandas as pd
 
-from sql_testing_library.adapters.redshift import RedshiftAdapter, RedshiftTypeConverter
-from sql_testing_library.mock_table import BaseMockTable
+from sql_testing_library._adapters.redshift import RedshiftAdapter, RedshiftTypeConverter
+from sql_testing_library._mock_table import BaseMockTable
 
 
 # Mock psycopg2 for testing
